requirements.txt
asgiref==3.4.1
attrs==23.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-crontab==0.7.1
django-decouple==2.1
django-environ==0.9.0
django-filter==21.1
django-graylog==0.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.3.0
PyPDF2==2.11.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0.1
referencing==0.35.1
requests==2.27.1
rpds-py==0.19.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
